Split up our Query augmentation into its own directory
