+++ src/core/database/query-augmentation/index.ts
@@ -1,0 +1,6 @@
+import './apply';
+import './as-result';
+import './call';
+import './log-it';
+import './map';
+import './subquery';
